+++ causallib/tests/test_positivity_metrics.py
@@ -1,6 +1,6 @@
 import unittest
 from causallib.positivity import Trimming
-from causallib.positivity.scorers import cross_covariance, cross_covariance_score
+from causallib.positivity.metrics import cross_covariance, cross_covariance_score
 from causallib.positivity.datasets.positivity_data_simulator import make_multivariate_normal_data
 from sklearn.linear_model import LogisticRegression
 import pandas as pd
